Renamed _core to _model_api.
